Refactored code into presenter and view packages + solve error
