+++ app/src/main/java/com/example/hsh/homesweethome/furniture/Furniture.java
@@ -2,7 +2,6 @@ package com.example.hsh.homesweethome.furniture;
 
 import android.support.annotation.NonNull;
 import android.support.design.widget.NavigationView;
-import android.support.v4.app.FragmentTransaction;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
@@ -12,8 +11,9 @@ import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 
-import com.arlib.floatingsearchview.FloatingSearchView;
 import com.example.hsh.homesweethome.R;
+import com.example.hsh.homesweethome.furniture.presenter.FurniturePresenter;
+import com.example.hsh.homesweethome.furniture.view.IFurnitureView;
 
 public class Furniture extends AppCompatActivity
         implements IFurnitureView, NavigationView.OnNavigationItemSelectedListener {
