+++ app/src/main/java/com/example/hsh/homesweethome/furniture/FurnitureMainFragment.java
@@ -3,32 +3,19 @@ package com.example.hsh.homesweethome.furniture;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.SearchView;
 
 import com.example.hsh.homesweethome.Models.CategoryFurniture;
-import com.example.hsh.homesweethome.Models.Furniture;
 import com.example.hsh.homesweethome.R;
-import com.example.hsh.homesweethome.network.APIService;
-import com.example.hsh.homesweethome.network.Utils;
-import com.jakewharton.rxbinding2.widget.TextViewTextChangeEvent;
+import com.example.hsh.homesweethome.furniture.presenter.FurnitureMainFragmentPresenter;
+import com.example.hsh.homesweethome.furniture.view.IFurnitureMainFragmentView;
 
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-
-import io.reactivex.disposables.CompositeDisposable;
-import io.reactivex.observers.DisposableObserver;
-import io.reactivex.subjects.PublishSubject;
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
 
 public class FurnitureMainFragment extends FurnitureBaseFragment implements IFurnitureMainFragmentView {
