--- app/src/main/java/com/example/hsh/homesweethome/furniture/IFragmentNavigationPresenter.java
@@ -1,5 +1,0 @@
-package com.example.hsh.homesweethome.furniture;
-
-public interface IFragmentNavigationPresenter {
-    void changeFragment(FurnitureBaseFragment fragment);
-}
