--- app/src/main/java/com/example/hsh/homesweethome/furniture/IFragmentNavigationView.java
@@ -1,5 +1,0 @@
-package com.example.hsh.homesweethome.furniture;
-
-public interface IFragmentNavigationView {
-    void attachPresenter(IFragmentNavigationPresenter presenter);
-}
