--- app/src/main/java/com/example/hsh/homesweethome/furniture/IFurnitureView.java
@@ -1,5 +1,0 @@
-package com.example.hsh.homesweethome.furniture;
-
-public interface IFurnitureView {
-    void setFragment(FurnitureBaseFragment fragment);
-}
