+++ app/src/main/java/com/example/hsh/homesweethome/furniture/presenter/IFragmentNavigationPresenter.java
@@ -1,0 +1,7 @@
+package com.example.hsh.homesweethome.furniture.presenter;
+
+import com.example.hsh.homesweethome.furniture.FurnitureBaseFragment;
+
+public interface IFragmentNavigationPresenter {
+    void changeFragment(FurnitureBaseFragment fragment);
+}
