+++ app/src/main/java/com/example/hsh/homesweethome/furniture/presenter/IFurnitureCategoryFragmentPresenter.java
@@ -1,4 +1,4 @@
-package com.example.hsh.homesweethome.furniture;
+package com.example.hsh.homesweethome.furniture.presenter;
 
 import android.os.Bundle;
