+++ app/src/main/java/com/example/hsh/homesweethome/furniture/presenter/IFurnitureMainFragmentPresenter.java
@@ -1,4 +1,4 @@
-package com.example.hsh.homesweethome.furniture;
+package com.example.hsh.homesweethome.furniture.presenter;
 
 public interface IFurnitureMainFragmentPresenter {
