+++ app/src/main/java/com/example/hsh/homesweethome/furniture/presenter/IFurniturePresenter.java
@@ -1,4 +1,4 @@
-package com.example.hsh.homesweethome.furniture;
+package com.example.hsh.homesweethome.furniture.presenter;
 
 public interface IFurniturePresenter {
