+++ app/src/main/java/com/example/hsh/homesweethome/furniture/presenter/IRecyclerViewAdapterMainPresenter.java
@@ -1,4 +1,6 @@
-package com.example.hsh.homesweethome.furniture;
+package com.example.hsh.homesweethome.furniture.presenter;
+
+import com.example.hsh.homesweethome.furniture.RecyclerViewHolderCategoryFurniture;
 
 public interface IRecyclerViewAdapterMainPresenter {
     void onBindFurnitureCategoryAtPosition(int position, RecyclerViewHolderCategoryFurniture viewHolder);
