+++ app/src/main/java/com/example/hsh/homesweethome/furniture/presenter/IRecyclerViewAdapterPresenter.java
@@ -1,4 +1,6 @@
-package com.example.hsh.homesweethome.furniture;
+package com.example.hsh.homesweethome.furniture.presenter;
+
+import com.example.hsh.homesweethome.furniture.RecyclerViewHolderFurniture;
 
 public interface IRecyclerViewAdapterPresenter {
     void onBindFurnitureViewAtPosition(int position, RecyclerViewHolderFurniture viewHolder);
