+++ app/src/main/java/com/example/hsh/homesweethome/furniture/view/IFurnitureMainFragmentView.java
@@ -1,4 +1,4 @@
-package com.example.hsh.homesweethome.furniture;
+package com.example.hsh.homesweethome.furniture.view;
 
 import com.example.hsh.homesweethome.Models.CategoryFurniture;
